backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.25.0
pydantic[email]==2.5.3
pydantic-settings==2.1.0
SQLAlchemy==2.0.23
asyncpg==0.29.0
psycopg[binary]==3.1.18
aiosqlite==0.19.0
alembic==1.13.1
python-multipart==0.0.6
python-dotenv==1.0.0
geoalchemy2==0.14.2
python-jose==3.3.0
passlib[bcrypt]==1.7.4
